requirements.txt
fastapi==0.81.0
loguru==0.6.0
python-docx==0.8.11
pymystem3==0.2.0
nltk==3.7
uvicorn==0.18.3
annoy~=1.17.1
navec~=0.10.0
transformers~=4.21.2
spacy~=3.1.0
pandas~=1.4.3
matplotlib~=3.5.3
torch~=1.12.1
tqdm~=4.64.0
scikit-learn~=1.1.2
pydantic~=1.8.2
python-multipart
textract